backend: Retrying watcher connection if dropped and double master check
The connection to etcd that watches for changes can drop during a node 
change. This is now retried. The container will still stop if the lease 
drops altogether, usually with a 10 second timeout.

There have also been cases where a second master has been registered 
which should never happen. This is being detected and the master 
promotion blocked.