Re-write of logic to handle losing the master lock and watch restart
The initial design for MSDHA expected that etcd locks would remain held
for the lifetime of the process run under the lock. It also expected the
process to be stopped and the lease to be dropped all at the same time
if the etcd lock was somehow lost. See
[here](https://github.com/etcd-io/etcd/issues/11457#issuecomment-574060952)
for more information.

The new design instead relies on detecting that there is no current
master rather than using a lock to hold the master role. MSDHA will (in
order):

  * wait MSDHA_TTL*2 seconds
  * Acquire the etcd lock
  * Determine if a master already exists by now otherwise stop
  * Promote the node
  * Set the state to master in etcd
  * Run a background job to wait MSDHA_MASTER_TTL seconds before
    terminating the container

The logic for restarting the etcd change detection was also re-written
to prevent losing events between restarts.